requirements.txt
pymongo>=4.8.0
PyYAML>=6.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
